429 fractional stock for the refill reminder (#438)
* Change database fields to double

* Proper formatting and conversion

* Rename .java to .kt

* Finish implementation

* Compile fix

* Update screenshots

* Allow flaky test

* Fix unit test

* Update documentation

* Fix refilling

* Allow flaky test

* Minor refactoring

+++ app/src/main/java/com/futsch1/medtimer/database/Medicine.java
@@ -38,13 +38,13 @@ public class Medicine {
     public OutOfStockReminderType outOfStockReminder;
     @ColumnInfo(defaultValue = "0")
     @Expose
-    public int amount;
+    public double amount;
     @ColumnInfo(defaultValue = "0")
     @Expose
-    public int outOfStockReminderThreshold;
+    public double outOfStockReminderThreshold;
     @ColumnInfo(defaultValue = "[]")
     @Expose
-    public ArrayList<Integer> refillSizes;
+    public ArrayList<Double> refillSizes;
 
     public Medicine(String name) {
         this.name = name;
